Add files via upload
Sentimental Analysis is a new variant in the research area. It basically refers to opinions or 
views of the different data that is being collected using surveys , comments and reviews 
over the web. The data set we have used in our project is from imdb movie rating website. 
Through this we can classify the review as positive or negative, a large amount of data is 
being generated daily. This will determine the polarity of reviews and these reviews will be 
classified as three types positive, negative and neutral.